cpp(io): satisfy clang-tidy (enum sizes, widening multiplications)
performance-enum-size: GgmlType and MapStage now use the smallest fitting
underlying type (uint8_t; codes/values stay <= 30). bugprone-implicit-
widening-of-multiplication-result: gguf_test.cpp's int*int products that feed
a size_t vector-size/offset now start with a size_t{...} operand.

+++ cpp/libs/sapient-io/include/sapient/io/gguf.hpp
@@ -115,7 +115,7 @@ inline constexpr uint32_t GGUF_MAGIC = 0x46554747; // "GGUF"
 inline constexpr uint64_t DEFAULT_ALIGNMENT = 32;
 inline constexpr size_t QK_K = 256;
 
-enum class GgmlType : uint32_t {
+enum class GgmlType : uint8_t {
     F32 = 0,
     F16 = 1,
     Q4_0 = 2,
